requirements.txt
# app2gcal dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Google API
google-api-python-client>=2.116.0
google-auth>=2.27.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
